Move the exported entry points to the main source file
This keeps `parser.ts` and `tokenizer.ts` only contain the respective
class and nothing else.

Fixes #1427
